refactor(doc-linker): lift DocAnchor + ANCHOR_* to leaf module (#501)
Move DocAnchor, ANCHOR_INGEST/MANUAL/DERIVED, ANCHOR_TYPES out of
doc_linker into the new leaf module aelfrice.doc_linker_types. Drop
the if-TYPE_CHECKING import of MemoryStore in doc_linker; the function
signatures keep the "MemoryStore" forward-ref string, which `from
__future__ import annotations` already makes a no-op at runtime.

doc_linker re-exports the lifted symbols so external callers
(`from aelfrice.doc_linker import DocAnchor`) keep working.

Mirrors the leaf-module pattern PR #500 used for np_pattern,
classification_core, and db_paths.

No behavior change. Sets up the store.py edge swap in the next commit.
